logo-NEO-Research-Peter-ECA
Hi guys,
The idea is to make a simple yet very recognisable logo for the Research dept within NEO. So why not just place a picture of a magnifying glass (which is genarally known as a symbol for (re)searching). I dont have the software to make the picture of the magnifying glass transparant, but the idea is to keep the NEO logo whole and place the transparant magnifying glass picture with proportional size within the square inside of the NEO logo.

Maybe you can do something with this idea. Maybe not. Just an idea when reading about it.

Kind regards,
Peter-ECA (also NEO HODL-er) ;)